Move class to package
S2-group/NAPPA#98
